xdiff: change rindex from long to size_t in xdfile_t
rindex describes a index offset which means it's an index into memory
which should use size_t. dstart and dend will be deleted in a future
patch series. Move them to the end to help avoid refactor conflicts.

+++ xdiff/xtypes.h
@@ -49,7 +49,7 @@ typedef struct s_xdfile {
 	xrecord_t *recs;
 	size_t nrec;
 	bool *changed;
-	long *rindex;
+	size_t *rindex;
 	size_t nreff;
 	ssize_t dstart, dend;
 } xdfile_t;
